chore: rename "queries" folder to "data"
