+++ src/data/mongo/index.ts
@@ -1,5 +1,5 @@
 import { type Db, connect } from '@rakered/mongo'
-import type { Item, ItemWithoutId, Queries } from './types'
+import type { Item, ItemWithoutId, Queries } from '../types'
 
 let uri: string
 let db: Db
